fix: tests & chore: move small csv export tests into router tests
