+++ backend/test/modules/location/location_router_test.py
@@ -3,16 +3,20 @@ import pytest
 import pytest_asyncio
 from httpx import AsyncClient
 from src.core.exceptions import InternalServerException
+from src.modules.incident.incident_model import IncidentSeverity
 from src.modules.location.location_entity import LocationEntity
 from src.modules.location.location_model import AutocompleteResult, LocationDto
 from src.modules.location.location_service import (
     LocationConflictException,
     LocationNotFoundException,
 )
+from test.modules.incident.incident_utils import IncidentTestUtils
 from test.modules.location.location_utils import GmapsMockUtils, LocationTestUtils
 from test.utils.http.assertions import assert_res_failure, assert_res_paginated, assert_res_success
 from test.utils.http.test_templates import (
+    assert_excel_response,
     generate_auth_required_tests,
+    generate_csv_empty_test,
     generate_filter_sort_tests,
     generate_search_tests,
 )
@@ -71,6 +75,16 @@ test_location_authentication = generate_auth_required_tests(
         "/api/locations/autocomplete",
         {"address": "123 Main St"},
     ),
+)
+
+test_location_csv_authentication = generate_auth_required_tests(
+    ({"admin", "staff"}, "GET", "/api/locations/csv", None),
+)
+
+test_location_csv_empty = generate_csv_empty_test(
+    "staff",
+    "/api/locations/csv",
+    ("Address", "Remote Warning Count", "In-Person Warning Count", "Citation Count"),
 )
 
 
@@ -340,7 +354,6 @@ class TestLocationCRUDRouter:
         data = assert_res_success(response, LocationDto)
 
         assert data.id == location.id
-        # When place_id unchanged, address data stays the same, only expiration can update
         assert data.google_place_id == location.google_place_id
         assert data.hold_expiration is None
 
@@ -387,7 +400,6 @@ class TestLocationCRUDRouter:
 
         self.location_utils.assert_matches(location, data)
 
-        # Verify deletion
         locations = await self.location_utils.get_all()
         assert len(locations) == 0
 
@@ -396,6 +408,71 @@ class TestLocationCRUDRouter:
         """Test deleting a non-existent location."""
         response = await self.admin_client.delete("/api/locations/999")
         assert_res_failure(response, LocationNotFoundException(999))
+
+
+class TestLocationCSVRouter:
+    """Tests for GET /api/locations/csv endpoint."""
+
+    staff_client: AsyncClient
+    location_utils: LocationTestUtils
+    incident_utils: IncidentTestUtils
+
+    @pytest.fixture(autouse=True)
+    def _setup(
+        self,
+        staff_client: AsyncClient,
+        location_utils: LocationTestUtils,
+        incident_utils: IncidentTestUtils,
+    ):
+        self.staff_client = staff_client
+        self.location_utils = location_utils
+        self.incident_utils = incident_utils
+
+    @pytest.mark.asyncio
+    async def test_get_locations_csv_with_data(self):
+        """Test Excel export with locations and incidents returns correct counts."""
+        location1 = await self.location_utils.create_one()
+        location2 = await self.location_utils.create_one()
+
+        await self.incident_utils.create_one(
+            location_id=location1.id, severity=IncidentSeverity.REMOTE_WARNING
+        )
+        await self.incident_utils.create_one(
+            location_id=location1.id, severity=IncidentSeverity.REMOTE_WARNING
+        )
+        await self.incident_utils.create_one(
+            location_id=location1.id, severity=IncidentSeverity.IN_PERSON_WARNING
+        )
+        await self.incident_utils.create_one(
+            location_id=location1.id, severity=IncidentSeverity.CITATION
+        )
+        await self.incident_utils.create_one(
+            location_id=location1.id, severity=IncidentSeverity.CITATION
+        )
+        await self.incident_utils.create_one(
+            location_id=location1.id, severity=IncidentSeverity.CITATION
+        )
+
+        response = await self.staff_client.get("/api/locations/csv")
+        rows = assert_excel_response(
+            response,
+            ("Address", "Remote Warning Count", "In-Person Warning Count", "Citation Count"),
+            expected_row_count=3,
+        )
+
+        data_rows = {row[0]: row for row in rows[1:]}
+
+        assert location1.formatted_address in data_rows
+        row1 = data_rows[location1.formatted_address]
+        assert row1[1] == 2
+        assert row1[2] == 1
+        assert row1[3] == 3
+
+        assert location2.formatted_address in data_rows
+        row2 = data_rows[location2.formatted_address]
+        assert row2[1] == 0
+        assert row2[2] == 0
+        assert row2[3] == 0
 
 
 class TestLocationAutocompleteRouter:
